Rename sites and areas route groups per terminology update
